requirements.txt
pypdf
python-dotenv
pathlib
jsonlines
langchain
langchain-huggingface
langdetect
weaviate-client
tqdm
requests
pandas